Topic subscriptions
commit 25e497ad0be3ff455e3442e54e3159b457675b34
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 25 14:50:28 2024 -0800

    Finished stats output in plain text

commit d584aec4f4cbb1703e84865590c10931681f84e7
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 25 01:14:45 2024 -0800

    Plain text output - WIP

commit 3f3bace7e67de6c7b5dd1a4514f70ad314e4a5b2
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 24 16:30:15 2024 -0800

    Finished stats but only in json frormat

commit fa28a3d1d1d8809d8c2b56361f8c09160088ac2b
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 24 12:07:59 2024 -0800

    Model overhaul, EntityRef etc

commit f00b8a2591ec887840920b0209b9620a233cf257
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sun Dec 22 13:19:58 2024 -0800

    Ledger stats

commit a6eeedaea8d70f82293e51a2dc7fc31435e55652
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sun Dec 22 12:12:05 2024 -0800

    Better ack/nack responses

commit c0f5503eeae016016cc61fedbc516797db574b38
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sun Dec 22 11:38:52 2024 -0800

    Clean up

commit 63fb6cddd135cd3382d9d8fe8f94c671e4201b75
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sat Dec 21 12:36:35 2024 -0800

    Consumer disconnect key-shared

commit 0e75ce46f69fc6cbf5471ef39dbc1a043d37d092
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Fri Dec 20 21:17:49 2024 -0800

    Key shared subscriptions - tested manually

commit 67c583f2b38146b4c27a49e8f2e92c249646712f
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Fri Dec 20 17:50:31 2024 -0800

    Key shared subscriptions - WIP

commit f9ae406029a04e44e9be1a8d140cb0ffed8f2f28
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Fri Dec 20 08:56:16 2024 -0800

    Dual key hash map wt=ithout arc

commit 5ff84f579086e1b0ab69d0876c79fa0003250d6e
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Thu Dec 19 19:01:42 2024 -0800

    Better version of DualKeyHashMap

commit 298fb29918d48d1d0d09e2b6f1d4985d73392420
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Thu Dec 19 17:59:53 2024 -0800

    DualKeyHashMap

commit 0c422654144c5aabdb65fff15ca826a5ef3e5ef1
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 18 22:02:44 2024 -0800

    Test client subscribes to messages

commit bbf796f0c1590c434e139712c016e983a5055579
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 18 21:17:53 2024 -0800

    Ack/Nack for shared subscriptions

commit d5197bdbdff9d86c1be90f6dab08f5e3a2a0ccba
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 18 14:22:10 2024 -0800

    Better response classification

commit b0c42b71f9a3e1494e62ccd16df1006a4e72738f
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Wed Dec 18 13:38:30 2024 -0800

    Subscriber get nest message

commit 40766b93d9a4530b7f82d9fdf5c7c68337dc1697
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 17 22:53:04 2024 -0800

    Allocating consumer ids

commit 97610d892f2faaa6e3f96ad24a9cb72feb885cfd
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 17 20:19:53 2024 -0800

    Subscriber API skeleton

commit 931a9577a9371ab6904596f2a3546d2a7a0eac80
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 17 17:59:50 2024 -0800

    Start of SubService

commit 221afb31aac691c211d2578e2be7d52e366dc59d
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 17 17:09:30 2024 -0800

    Subscriptions WIP

commit dcec20392f7ba1f7256635b14d7f93445175cffd
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Tue Dec 17 10:52:24 2024 -0800

    Adding messages to subscribers

commit aca2a499e8a33be6cd45dbc27ec02ea4f00ed728
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Mon Dec 16 22:42:00 2024 -0800

    WIP

commit 28d0eec7e39e38a355daa605a51d5ce79a14101f
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Mon Dec 16 11:14:21 2024 -0800

    WIP

commit f8481704de62e6a9d2cc58d121b92c1bd36c1e3d
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sun Dec 15 22:26:03 2024 -0800

    WIP

commit 2cc615f0403bba487fd77573c73980b1a58edf4f
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sun Dec 15 22:00:53 2024 -0800

    WIP

commit f255c2db07aaa501106895a13f0e56728f2f1858
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sat Dec 14 11:46:19 2024 -0800

    Log layout

commit fe2c62e0e1e8a375ed4ab2615b892144431b506f
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sat Dec 14 11:28:15 2024 -0800

    More data and better logs

commit a98360e944c40e4bce4a2fe3104a1055a6434853
Author: Martin Halliday <martin.halliday@narvar.com>
Date:   Sat Dec 14 09:35:37 2024 -0800

    New consumer event types

+++ net/src/data_types.rs
@@ -2,14 +2,14 @@
 Aliases for data types for readability and to allow resizing in future
 */
 
-pub type Timestamp = u64;
-pub type VersionNumber = u32;
-pub type PortNumber = u16;
+pub type Timestamp = u64; // Compatible with Rust time library
+pub type VersionNumber = u32; // Allows 4 billion updates to each database record
+pub type PortNumber = u16; // Conforms to TCP/IP port numbering
 
-pub type NodeId = u16;
-pub type TopicId = u32;
-pub type PartitionId = u16;
-pub type LedgerId = u32;
-pub type MessageId = u32;
-pub type SubscriptionId = u32;
-pub type ConsumerId = u128;
+pub type NodeId = u16; // Maximum of 65 thousand nodes in a cluster
+pub type TopicId = u32; // Up to 4 billion topics per cluster
+pub type PartitionId = u16; // Up to 65 thousand partitions per topic
+pub type LedgerId = u32; // Up to 4 billion ledgers per partition
+pub type MessageId = u32; // // Up to 4 billion messages per ledger
+pub type SubscriptionId = u32; // // Up to 4 billion subscriptions per topic
+pub type ConsumerId = u64; // Allows us to create 1 million consumers per second for about half a million years
